Improve provider generator
